Implement MoShuffleboard
MoShuffleboard is an attempt to emulate the ease of the SmartDashboard
put* api, while still leaving room for layout customization. The goal is to
create an interface with the SmartDashboard that is both simple and powerful.

This is accomplished by allowing the user to defer layout considerations.
Essentially, the MoShuffleboard put* apis allow the programmer to
immediately write code that pushes values to the Shuffleboard, then
write seperate code to determine the layout and position of the pushed values
at a later date.

This patch has two main benefits:
- All shuffleboard layout code is gathered into one place. This makes
  updating layouts much easier as the programmer no longer has to jump
  around the various files in order to figure out what is pushing which
  values to which grid locations.
- Enables quick iteration, with a later clean-up pass. This means that a
  programmer can decide to put a value on the Shuffleboard, then defer
  the layout considerations for a future date. This should help enable
  the rapid development required during competitions, since a value
  can quickly be pushed to the Shuffleboard, then during post-comp cleanup
  the layout position can be reconsidered and improved.