[Climber] Add current detection for zeroing
As discussed in the meeting today, we can use current limiting to detect
when the lift has reached its zero. This patch adds that capability, just
in case we don't get the limit switches installed in time for Beach Blitz.

This feature can be disabled via a boolean preference in MoPrefs, which
will cause the climber to fall back to zeroing with limit switches.