Remove motor control from ShooterSubsystem::periodic
Motor control should always be done from within a command, so that the
scheduler can guarantee that there are no motor conflicts.

For example, if a command were to try to do one thing with a motor, and
the subystem::periodic were trying to do a different thing, we could get
unpredictable behavior. To prevent this, we only ever set motors from
within commands which have reserved that subsystem. That way the scheduler
will guarantee no two commands using the same subystem run at the same
time.